+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useInputId } from './use-input-id'
+export { default as useForm } from './use-form'