+++ CHANGELOG.md
@@ -11,3 +11,5 @@
 -   [added]: TextField component
 -   [added]: Icon support
 -   [added]: Support for icon in buttons and inputs.
+-   [added]: Control for the icon in the button, end or lead
+-   [added]: useForm hook